move compatibility stuff to compat.py
