+++ gearman/protocol.py
@@ -1,7 +1,7 @@
 import struct
 from gearman.constants import PRIORITY_NONE, PRIORITY_LOW, PRIORITY_HIGH
 from gearman.errors import ProtocolError
-from gearman.util import any,all
+from gearman import compat
 # Protocol specific constants
 NULL_CHAR = '\x00'
 MAGIC_RES_STRING = '%sRES' % NULL_CHAR
@@ -238,7 +238,7 @@ def pack_binary_command(cmd_type, cmd_args, is_response=False):
 
     # !NOTE! str should be replaced with bytes in Python 3.x
     # We will iterate in ORDER and str all our command arguments
-    if any(type(param_value) != str for param_value in cmd_args.itervalues()):
+    if compat.any(type(param_value) != str for param_value in cmd_args.itervalues()):
         raise ProtocolError('Received non-binary arguments: %r' % cmd_args)
 
     data_items = [cmd_args[param] for param in expected_cmd_params]
