+++ tests/worker_tests.py
@@ -1,6 +1,5 @@
-from collections import deque
-try: from collections import defaultdict
-except ImportError: from gearman.util import defaultdict
+import collections
+from gearman import compat
 import unittest
 
 from gearman.worker import GearmanWorker
@@ -16,7 +15,7 @@ from tests._core_testing import _GearmanAbstractTest, MockGearmanConnectionManag
 class MockGearmanWorker(MockGearmanConnectionManager, GearmanWorker):
     def __init__(self, *largs, **kwargs):
         super(MockGearmanWorker, self).__init__(*largs, **kwargs)
-        self.worker_job_queues = defaultdict(deque)
+        self.worker_job_queues = compat.defaultdict(collections.deque)
 
     def on_job_execute(self, current_job):
         current_handler = self.connection_to_handler_map[current_job.connection]
